[167][Clean_Auth] Auth42 Clean in auth.js
