+++ front/app/src/views/view-login.ce.js
@@ -3,6 +3,7 @@ import { redirectTo } from '@/router.js';
 import { user } from '@/auth.js';
 import { getCsrfToken } from '@/auth.js';
 import { loginUser } from '@/auth.js';
+import { getAuthorizationCode } from '@/auth.js';
 
 class ViewSignIn extends HTMLElement {
   connectedCallback() {
@@ -76,7 +77,7 @@ class ViewSignIn extends HTMLElement {
 
     this.querySelector('#OAuth-42').addEventListener('click', event => {
       event.preventDefault();
-      this.getAuthorizationCode();
+      getAuthorizationCode();
     });
   }
 
@@ -114,19 +115,6 @@ class ViewSignIn extends HTMLElement {
         console.error('Login failed:', error);
     }
   }
-
-  getAuthorizationCode() {
-    const authorizationUrl =
-        "https://api.intra.42.fr/oauth/authorize";
-    const clientId =
-        "u-s4t2ud-032700fdff8bf6b743669184234c5670698f0f0ef95b498514fc13b5e7af32f0";
-    const redirectUri =
-        "https%3A%2F%2F127.0.0.1%3A5500%2FWeb%2Fbackend%2Fauthentification%2Ftemplates%2Flogin_with42api.html";
-    const responseType = "code";
-    // const url = `${authorizationUrl}?client_id=${clientId}&redirect_uri=${redirectUri}&response_type=${responseType}`;
-    const url = `https://api.intra.42.fr/oauth/authorize?client_id=u-s4t2ud-032700fdff8bf6b743669184234c5670698f0f0ef95b498514fc13b5e7af32f0&redirect_uri=http%3A%2F%2F127.0.0.1%3A8000%2F&response_type=code`;
-    window.location.href = url;
-  }  
 }
 
 customElements.define('view-signin', ViewSignIn);
