+++ front/app/src/views/view-signup.ce.js
@@ -2,6 +2,7 @@ import '@/components/layouts/auth-layout/auth-layout.ce.js';
 import { redirectTo } from '@/router.js';
 import { getCsrfToken } from '@/auth.js';
 import { sendSignUpRequest } from '@/auth.js';
+import { getAuthorizationCode } from '@/auth.js';
 
 class ViewSigUp extends HTMLElement {
   constructor() {
@@ -119,7 +120,7 @@ class ViewSigUp extends HTMLElement {
 
     this.querySelector('#OAuth-42').addEventListener('click', event => {
       event.preventDefault();
-      this.getAuthorizationCode();
+      getAuthorizationCode();
     });
   }
 
@@ -205,19 +206,6 @@ class ViewSigUp extends HTMLElement {
     this.usernameError.textContent = '';
     this.passwordError.textContent = '';
   };
-
-  getAuthorizationCode() {
-    const authorizationUrl =
-        "https://api.intra.42.fr/oauth/authorize";
-    const clientId =
-        "u-s4t2ud-032700fdff8bf6b743669184234c5670698f0f0ef95b498514fc13b5e7af32f0";
-    const redirectUri =
-        "https%3A%2F%2F127.0.0.1%3A5500%2FWeb%2Fbackend%2Fauthentification%2Ftemplates%2Flogin_with42api.html";
-    const responseType = "code";
-    // const url = `${authorizationUrl}?client_id=${clientId}&redirect_uri=${redirectUri}&response_type=${responseType}`;
-    const url = `https://api.intra.42.fr/oauth/authorize?client_id=u-s4t2ud-032700fdff8bf6b743669184234c5670698f0f0ef95b498514fc13b5e7af32f0&redirect_uri=http%3A%2F%2F127.0.0.1%3A8000%2F&response_type=code`;
-    window.location.href = url;
-}
 }
 
 customElements.define('view-signup', ViewSigUp);
